Rename shopping cart reducer file to remove hook wording
